pyproject.toml
[project]
name = "pinch-core"
version = "0.1.0"
description = "Pinch of Salt news processing engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "requests>=2.31.0",
    "scikit-learn>=1.4.0",
    "pydantic>=2.6.0",
    "instructor>=1.0.0",
    "python-dotenv>=1.0.1",
    "uuid>=1.30",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "numpy>=2.4.2",
]

[tool.uv]
managed = true
